FIXED the output of SUBPROCESS
I needed to use the 'Popen' function within subprocess library because the original 'run' function would only print CompletedProcess(), with all the data, but we just wanted to use the stdout data. We also ran into a problem were the output of Popen was returning 'None' reading the python docs, it says we must also enable stdout=PIPE and/or stderr=PIPE, to remove this return type, so we did add them to the code and now it works! Nmap prints out the data within our box. I'm happy:)